Load manifest only once
