+++ tests/test_update.py
@@ -34,7 +34,7 @@ def test_update(name, projects):
 
     with patch("dfetch.commands.update.SuperProject", return_value=fake_superproject):
         with patch(
-            "dfetch.manifest.manifest.get_childmanifests"
+            "dfetch.manifest.parse.get_childmanifests"
         ) as mocked_get_childmanifests:
             with patch("dfetch.project.make") as mocked_make:
                 with patch("os.path.exists"):
@@ -58,7 +58,7 @@ def test_forced_update():
 
     with patch("dfetch.commands.update.SuperProject", return_value=fake_superproject):
         with patch(
-            "dfetch.manifest.manifest.get_childmanifests"
+            "dfetch.manifest.parse.get_childmanifests"
         ) as mocked_get_childmanifests:
             with patch("dfetch.project.make") as mocked_make:
                 with patch("os.path.exists"):
